Added logout functionality, auth & guest middleware to routes and modified views

+++ routes/web.php
@@ -21,40 +21,42 @@ Route::get('/', [ProductController::class, 'index'])->name('index');
 
 Route::prefix('/products')->group(function () {
     // Route to get form to create a product
-    Route::get('/create', [ProductController::class, 'create'])->name('create');
+    Route::get('/create', [ProductController::class, 'create'])->name('create')->middleware('auth');
 
     // Route to get categories
-    Route::get('/all', [PageController::class, 'all'])->name('all');
+    Route::get('/all', [PageController::class, 'all'])->name('all')->middleware('auth');
 
     // Route to get a single product
-    Route::get('/show/{product}', [ProductController::class, 'show'])->name('show');
+    Route::get('/show/{product}', [ProductController::class, 'show'])->name('show')->middleware('auth');
 
     // Route to store a product
-    Route::post('/', [ProductController::class, 'store'])->name('store');
+    Route::post('/', [ProductController::class, 'store'])->name('store')->middleware('auth');
 
     // Route to get form to update a product
-    Route::get('/update/{product}', [ProductController::class, 'edit'])->name('edit');
+    Route::get('/update/{product}', [ProductController::class, 'edit'])->name('edit')->middleware('auth');
 
     // Route to update a product
-    Route::put('/update/{product}', [ProductController::class, 'update'])->name('update');
+    Route::put('/update/{product}', [ProductController::class, 'update'])->name('update')->middleware('auth');
 
     // Route to delete a product
-    Route::delete('/{product}', [ProductController::class, 'destroy'])->name('delete');
+    Route::delete('/{product}', [ProductController::class, 'destroy'])->name('delete')->middleware('auth');
 });
 
 
 Route::prefix('/users')->group(function() {
     // Route to get login page
-    Route::get('/login', [AuthController::class, 'login'])->name('login');
+    Route::get('/login', [AuthController::class, 'login'])->name('login')->middleware('guest');
     
-    // Route to login user
+    // Route to login a user
     Route::post('/login', [AuthController::class, 'authenticate'])->name('user.auth');
 
     // Route to get register page
-    Route::get('/register', [AuthController::class, 'register'])->name('register');
+    Route::get('/register', [AuthController::class, 'register'])->name('register')->middleware('guest');
     
-    // Route to create user
+    // Route to create a user
     Route::post('/register', [AuthController::class, 'store'])->name('user.store');
 
+    // Route to logout a user
+    Route::post('/logout', [AuthController::class, 'logout'])->name('logout')->middleware('auth');
 });
